# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "超感覚C3-2023 API",
        "contact": {},
        "license": {
            "name": "Simo-C3"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/todo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ToDo"
                ],
                "summary": "Todo一覧を配列で返す",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ToDos"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ToDo"
                ],
                "summary": "ToDoを作成する",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PostToDo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ToDo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/todo/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ToDo"
                ],
                "summary": "ToDoを編集する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ToDo"
                ],
                "summary": "ToDoを編集する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PatchToDo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ToDo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BaseStatus": {
            "type": "string",
            "enum": [
                "todo",
                "doing",
                "done"
            ],
            "x-enum-varnames": [
                "Todo",
                "Doing",
                "Done"
            ]
        },
        "models.ToDo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.BaseStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "schemas.DeleteSuccessResponse": {
            "type": "string",
            "enum": [
                "success"
            ],
            "x-enum-varnames": [
                "DeleteStatus"
            ]
        },
        "schemas.PatchToDo": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.BaseStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schemas.PostToDo": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.BaseStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schemas.ToDo": {
            "type": "object",
            "properties": {
                "todo": {
                    "$ref": "#/definitions/models.ToDo"
                }
            }
        },
        "schemas.ToDos": {
            "type": "object",
            "properties": {
                "todos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ToDo"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}